+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add BESSI Survey [#2491](https://github.com/rokwire/illinois-app/issues/2491)
 
 ## [4.2.30] - 2022-11-30
 ### Changed